--- src/main/java/presentation/panels/addentitypanels/AddAnimalPanel.java
@@ -1,10 +1,0 @@
-package presentation.panels.addentitypanels;
-
-import presentation.frames.MainFrame;
-
-public class AddAnimalPanel extends AddEntityPanel{
-
-    public AddAnimalPanel(MainFrame owner) {
-        super(owner);
-    }
-}
