--- src/main/java/presentation/panels/addentitypanels/AddShelterPanel.java
@@ -1,10 +1,0 @@
-package presentation.panels.addentitypanels;
-
-import presentation.frames.MainFrame;
-
-public class AddShelterPanel extends AddEntityPanel{
-
-    public AddShelterPanel(MainFrame owner) {
-        super(owner);
-    }
-}
