--- src/main/java/presentation/panels/addentitypanels/AddVolunteerPanel.java
@@ -1,10 +1,0 @@
-package presentation.panels.addentitypanels;
-
-import presentation.frames.MainFrame;
-
-public class AddVolunteerPanel extends AddEntityPanel{
-
-    public AddVolunteerPanel(MainFrame owner) {
-        super(owner);
-    }
-}
