--- src/main/java/presentation/styles/Style.java
@@ -1,17 +1,0 @@
-package presentation.styles;
-
-import java.awt.*;
-
-public class Style {
-    public static final Color COLOR_BACKGROUND = new Color(237, 237, 237);
-    public static final Color COLOR_BACKGROUND_DARK = new Color(208, 208, 208);
-    public static final Color COLOR_BTN = new Color(42, 122, 253);
-    public static final Color COLOR_BTN_HOVER = new Color(80, 144, 248);
-    public static final Color COLOR_BTN_BACK = new Color(44, 44, 44);
-    public static final Color COLOR_BTN_BACK_HOVER = new Color(60, 60, 60);
-    public static final Color COLOR_BTN_DELETE = new Color(253, 94, 86);
-    public static final Color COLOR_BTN_DELETE_HOVER = new Color(251, 115, 108);
-    public static final Color COLOR_STATE_PENDING = new Color(253, 202, 0);
-    public static final Color COLOR_STATE_CANCELED = new Color(253, 56, 60);
-
-}
