+++ src/main/java/views/panels/addentitypanels/AddAnimalPanel.java
@@ -1,0 +1,10 @@
+package views.panels.addentitypanels;
+
+import views.frames.MainFrame;
+
+public class AddAnimalPanel extends AddEntityPanel{
+
+    public AddAnimalPanel(MainFrame owner) {
+        super(owner);
+    }
+}
