+++ src/main/java/views/panels/addentitypanels/AddShelterPanel.java
@@ -1,0 +1,10 @@
+package views.panels.addentitypanels;
+
+import views.frames.MainFrame;
+
+public class AddShelterPanel extends AddEntityPanel{
+
+    public AddShelterPanel(MainFrame owner) {
+        super(owner);
+    }
+}
