+++ src/main/java/views/panels/addentitypanels/AddVolunteerPanel.java
@@ -1,0 +1,10 @@
+package views.panels.addentitypanels;
+
+import views.frames.MainFrame;
+
+public class AddVolunteerPanel extends AddEntityPanel{
+
+    public AddVolunteerPanel(MainFrame owner) {
+        super(owner);
+    }
+}
